+++ src/Application/_Env/NodeJS/Cookie.ts
@@ -1,5 +1,6 @@
 /// <amd-module name="Application/_Env/NodeJS/Cookie" />
-import { IHttpRequest, IHttpResponse } from 'Application/Interface';
+import type { IHttpRequest } from 'Application/_Env/IHttpRequest';
+import type { IHttpResponse } from 'Application/_Env/IHttpResponse';
 import type { ICookie, ICookieOptions } from 'Application/_Env/Interfaces';
 
 /**
